+++ packages/preact_signals/lib/src/computed.dart
@@ -159,9 +159,12 @@ class Computed<T> with Listenable, ReadonlySignal<T> {
   }
 
   @override
+  @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
+  @pragma('wasm:prefer-inline')
   T get value {
     if ((flags & RUNNING) != 0) {
-      throw Exception('Cycle detected');
+      throwCycleDetected();
     }
 
     final node = addDependency();
@@ -170,7 +173,7 @@ class Computed<T> with Listenable, ReadonlySignal<T> {
       node.version = version;
     }
     if ((flags & HAS_ERROR) != 0) {
-      throw error!;
+      throwSignalEffectException(error!);
     }
     return _internalValue;
   }