+++ packages/preact_signals/lib/src/effect.dart
@@ -44,6 +44,9 @@ class Effect with Listenable {
         name = options?.name;
 
   @internal
+  @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
+  @pragma('wasm:prefer-inline')
   void callback() {
     final finish = start();
     try {
@@ -61,9 +64,12 @@ class Effect with Listenable {
   }
 
   @internal
+  @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
+  @pragma('wasm:prefer-inline')
   void Function() start() {
     if ((flags & RUNNING) != 0) {
-      throw Exception('Cycle detected');
+      throwCycleDetected();
     }
     flags |= RUNNING;
     flags &= ~DISPOSED;
@@ -77,6 +83,9 @@ class Effect with Listenable {
   }
 
   @override
+  @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
+  @pragma('wasm:prefer-inline')
   void notify() {
     if (!((flags & NOTIFIED) != 0)) {
       flags |= NOTIFIED;
@@ -148,7 +157,7 @@ class Effect with Listenable {
   void endEffect(Listenable? prevContext) {
     final effect = this;
     if (evalContext != effect) {
-      throw Exception('Out-of-order effect');
+      throwOutOfOrderEffect();
     }
     effect.cleanupSources();
     evalContext = prevContext;