+++ packages/signals_hooks/lib/src/async.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:signals_flutter/signals_core.dart';
+import 'package:signals/signals_core.dart';
 
 import 'core.dart';