+++ packages/signals_hooks/lib/src/core.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:signals_flutter/signals_core.dart';
+import 'package:signals/signals_flutter.dart';
 
 import 'base.dart';
 
@@ -67,7 +67,7 @@ T useSignalValue<T, S extends ReadonlySignal<T>>(
 ///   }
 /// }
 /// ```
-Signal<T> useSignal<T>(
+FlutterSignal<T> useSignal<T>(
   T value, {
   /// A list of objects to watch for changes.
   ///
@@ -95,7 +95,7 @@ Signal<T> useSignal<T>(
 ///   }
 /// }
 /// ```
-Computed<T> useComputed<T>(
+FlutterComputed<T> useComputed<T>(
   T Function() value, {
   /// A list of objects to watch for changes.
   ///